abstracted utils into various categories
